+++ lib/internal/bootstrap/node.js
@@ -114,24 +114,6 @@ if (isMainThread) {
   process.exit = wrapped.exit;
 }
 
-const {
-  emitWarning
-} = NativeModule.require('internal/process/warning');
-
-process.emitWarning = emitWarning;
-
-const {
-  nextTick,
-  runNextTicks
-} = NativeModule.require('internal/process/task_queues').setupTaskQueue();
-
-process.nextTick = nextTick;
-// Used to emulate a tick manually in the JS land.
-// A better name for this function would be `runNextTicks` but
-// it has been exposed to the process object so we keep this legacy name
-// TODO(joyeecheung): either remove it or make it public
-process._tickCallback = runNextTicks;
-
 const credentials = internalBinding('credentials');
 if (credentials.implementsPosixCredentials) {
   process.getuid = credentials.getuid;
@@ -272,6 +254,39 @@ Object.defineProperty(process, 'features', {
     setUncaughtExceptionCaptureCallback;
   process.hasUncaughtExceptionCaptureCallback =
     hasUncaughtExceptionCaptureCallback;
+}
+
+const {
+  emitWarning
+} = NativeModule.require('internal/process/warning');
+
+process.emitWarning = emitWarning;
+
+// We initialize the tick callbacks and the timer callbacks last during
+// bootstrap to make sure that any operation done before this are synchronous.
+// If any ticks or timers are scheduled before this they are unlikely to work.
+{
+  const {
+    nextTick,
+    runNextTicks
+  } = NativeModule.require('internal/process/task_queues').setupTaskQueue();
+
+  process.nextTick = nextTick;
+  // Used to emulate a tick manually in the JS land.
+  // A better name for this function would be `runNextTicks` but
+  // it has been exposed to the process object so we keep this legacy name
+  // TODO(joyeecheung): either remove it or make it public
+  process._tickCallback = runNextTicks;
+
+  const { getTimerCallbacks } = NativeModule.require('internal/timers');
+  const { setupTimers } = internalBinding('timers');
+  const { processImmediate, processTimers } = getTimerCallbacks(runNextTicks);
+  // Sets two per-Environment callbacks that will be run from libuv:
+  // - processImmediate will be run in the callback of the per-Environment
+  //   check handle.
+  // - processTimers will be run in the callback of the per-Environment timer.
+  setupTimers(processImmediate, processTimers);
+  // Note: only after this point are the timers effective
 }
 
 function setupProcessObject() {
